Remove dummy var and insert real label mikropml
Using dummy var in mikropml was one-hot encoding any factor that would be used as label later one. Weirdly, ranger does not like 0,1 for labels. So functionality now is to specify the factor of interest. If it detects the factor is a character, it will set it aside so it doesnt get one-hot encoded.