Big changes see desc
So...its a real problem when people copy column names across datasets but they are not the same data. What. The. Heck. I try and address this - well for numeric data...not so well for character data. I also try and make a robust way to deal with labels that are copied across datasets. We'll see what happens when the next person breaks it.