Increase shap sim and change adjust = T
So apparently, fastshap does "approximate" shap values. This does not satisfy the efficency property of shap values, that is the feature contributions have to add up to the difference between the prediction for x and the average training prediction. This means that the shap values are not fairly distributed between features. This can be approximated/converged on by increasing nsim. I'm doing both adjust = T (we can use a regression-based adjustment to correct the sum) and increasing nsim here.